PM-15155 - Move 2FA setup components into own folder
