+++ apps/web/src/app/auth/settings/security/security-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
 import { ChangePasswordComponent } from "../change-password.component";
-import { TwoFactorSetupComponent } from "../two-factor-setup.component";
+import { TwoFactorSetupComponent } from "../two-factor-setup/two-factor-setup.component";
 
 import { SecurityKeysComponent } from "./security-keys.component";
 import { SecurityComponent } from "./security.component";
