+++ apps/web/src/app/auth/settings/two-factor-setup/two-factor-authenticator.component.html
