+++ apps/web/src/app/auth/settings/two-factor-setup/two-factor-duo.component.ts
