+++ apps/web/src/app/auth/settings/two-factor-setup/two-factor-setup.component.ts
