+++ apps/web/src/app/shared/loose-components.module.ts
@@ -41,14 +41,14 @@ import { ApiKeyComponent } from "../auth/settings/security/api-key.component";
 import { ChangeKdfModule } from "../auth/settings/security/change-kdf/change-kdf.module";
 import { SecurityKeysComponent } from "../auth/settings/security/security-keys.component";
 import { SecurityComponent } from "../auth/settings/security/security.component";
-import { TwoFactorAuthenticatorComponent } from "../auth/settings/two-factor-authenticator.component";
-import { TwoFactorDuoComponent } from "../auth/settings/two-factor-duo.component";
-import { TwoFactorEmailComponent } from "../auth/settings/two-factor-email.component";
-import { TwoFactorRecoveryComponent } from "../auth/settings/two-factor-recovery.component";
-import { TwoFactorSetupComponent } from "../auth/settings/two-factor-setup.component";
-import { TwoFactorVerifyComponent } from "../auth/settings/two-factor-verify.component";
-import { TwoFactorWebAuthnComponent } from "../auth/settings/two-factor-webauthn.component";
-import { TwoFactorYubiKeyComponent } from "../auth/settings/two-factor-yubikey.component";
+import { TwoFactorAuthenticatorComponent } from "../auth/settings/two-factor-setup/two-factor-authenticator.component";
+import { TwoFactorDuoComponent } from "../auth/settings/two-factor-setup/two-factor-duo.component";
+import { TwoFactorEmailComponent } from "../auth/settings/two-factor-setup/two-factor-email.component";
+import { TwoFactorRecoveryComponent } from "../auth/settings/two-factor-setup/two-factor-recovery.component";
+import { TwoFactorSetupComponent } from "../auth/settings/two-factor-setup/two-factor-setup.component";
+import { TwoFactorVerifyComponent } from "../auth/settings/two-factor-setup/two-factor-verify.component";
+import { TwoFactorWebAuthnComponent } from "../auth/settings/two-factor-setup/two-factor-webauthn.component";
+import { TwoFactorYubiKeyComponent } from "../auth/settings/two-factor-setup/two-factor-yubikey.component";
 import { UserVerificationModule } from "../auth/shared/components/user-verification";
 import { SsoComponent } from "../auth/sso.component";
 import { TwoFactorOptionsComponent } from "../auth/two-factor-options.component";
